fix(cli): update output format for orchestrate agents list -v to be valid JSON without newlines(#1488)
here is an issue when running orchestrate agents list -v where the outputted text has newlines included in it, even when piped directly into pbcopy e.g. orchestrate agents list -v | pbcopy. As far as I can tell this is due to rich.print's behaviour to detect the console width and automatically wrap, breaking the validity of the JSON output.

The change in this PR is to make use of rich's print_json functionality which doesn't exhibit the same behaviour, while still looking the same to an end-user who is viewing in the console. This allows the output of orchestrate agents list -v to now be chained into other tools to process and act upon it.

Old behaviour output (truncated for readability)

[
  {
    "kind": "native",
    "id": "07cfd127-da08-4319-89fc-5f5f529f6da1",
    "name": "product_interaction_agent",
    "display_name": "product_interaction_agent",
    "description": "You are an agent that will gather data from the database and answer 
the users query about customer interactions for particu;ar store zones\n","context_access_enabled": true,
    "context_variables": [],
New behaviour:

[
  {
    "kind": "native",
    "id": "07cfd127-da08-4319-89fc-5f5f529f6da1",
    "name": "product_interaction_agent",
    "display_name": "product_interaction_agent",
    "description": "You are an agent that will gather data from the database and answer the users query about customer interactions for particu;ar store zones\n",
    "context_access_enabled": true,