feat(flows): Updates procode flows to support data mapping, genai nodes, foreach, and user activity (#1693)
Merge tempus branch in 1.11

What's in this PR:

there is a number of samples being added
Data mapping - we added support to specify explicit data mapping in a flow node
GenAI node - we added the ability to enable or disable examples provided in a genAI node
For Each - we enabled configuration of a for-each node to execute iterations in parallel
User Activities - enabled modeling of user activities using procode
All of these changes are localized to the flow builder. There are no changes to the ADK foundation code.