fix(groq): Fix default llm configuration (#2244)
The previous implementation made it so groq/openai/gpt-oss-120b was always set as the DEFAULT_LLM_MODEL. This is because setdefault does not set a value in a dictionary if it already exists and the DEFAULT_LLM_MODEL is set in the default.env meaning it was always set so it never got changed.

This change fixes this be directly overwriting the value. Though, it has a caveat that is the value was present in the user's env then do not overwrite it. This allows users to still set these values in their .env files if they want to explicitly set them